+++ CHANGES.rst
@@ -11,6 +11,8 @@ Development
 - [mw] Fix tests
 - [mw] Resolve issue when European publication server returns
   reference to WIPO as HTML response instead of PDF document
+- [mw] Pad document numbers for granted patents to 8 digits
+  with leading zeros when accessing USPTO for PDF documents
 
 
 2019-02-21 0.163.0